+++ src/om_inputs/schemas.cljs
@@ -54,5 +54,6 @@
               (s/optional-key :labeled)   s/Bool
               (s/optional-key :attrs)     s/Any
               (s/optional-key :className) s/Str
-              (s/optional-key :layout)    s/Str}})
+              (s/optional-key :layout)    s/Str
+              (s/optional-key :label-order)   s/Bool}})