+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * ADD new input diplay : stepper -/+ for s/Int
 * ADD Possibility to put an HTML description to a field with sablono.
 * ADD HTML Attributes for action and clean buttons. {:action {:attrs {:tabIndex 10}}
+* ADD For s/enum add the possibility to iterate over the i18n for order reason : {:label-order true} Only i18n data are shwon.
+* TECH upgrade to schema 0.3.3. Add s/set-compile-fn-validation! to false.
 
 ## 0.3.6-SNAPSHOT